nasa-basins.

+++ nasa_basins.f90
@@ -71,8 +71,8 @@ program nasa_basins
     end if 
 
     ! Determine how many lon and lat values will be output
-    nlon = ceiling( (maxval(inb%lon)-minval(inb%lon)) / dlon )
-    nlat = ceiling( (maxval(inb%lat)-minval(inb%lat)) / dlat )
+    nlon = int( ceiling( (maxval(inb%lon)-minval(inb%lon)) / dlon ) )
+    nlat = int( ceiling( (maxval(inb%lat)-minval(inb%lat)) / dlat ) )
     npo  = nlon*nlat 
 
     ! Allocate output vectors
